refactor: Add suggested changes and move AsyncTransport definition
