+++ engineio/src/transports/websocket.rs
@@ -1,7 +1,9 @@
-use crate::error::Result;
-use crate::header::HeaderMap;
-use crate::transport::{AsyncTransport, Transport};
-use crate::transports_async::websocket::AsyncWebsocketTransport;
+use crate::{
+    async_transports::{transport::AsyncTransport, websocket::AsyncWebsocketTransport},
+    error::Result,
+    header::HeaderMap,
+    transport::Transport,
+};
 use bytes::Bytes;
 use std::sync::Arc;
 use tokio::runtime::Runtime;
@@ -26,7 +28,7 @@ impl WebsocketTransport {
 
         let mut req = http::Request::builder().uri(url.clone().as_str());
         if let Some(map) = headers {
-            // SAFETY: this unwrap never panics as the underlying request is always in proper state
+            // SAFETY: this unwrap never panics as the underlying request is just initialized and in proper state
             req.headers_mut()
                 .unwrap()
                 .extend::<reqwest::header::HeaderMap>(map.try_into()?);
